+++ packages/kit/src/core/sync/create_manifest_data/index.js
@@ -482,18 +482,17 @@ function create_remotes(config, cwd) {
 
 	const externals = config.kit.remoteFunctions.allowedPaths.map((dir) => path.resolve(dir));
 
-	for (const dir of [config.kit.files.lib, config.kit.files.routes, ...externals]) {
+	for (const dir of [config.kit.files.src, ...externals]) {
 		if (!fs.existsSync(dir)) continue;
 
 		for (const file of walk(dir)) {
 			if (extensions.some((ext) => file.endsWith(ext))) {
 				const posixified = posixify(path.relative(cwd, `${dir}/${file}`));
 
-				remotes.push({
-					hash: hash(posixified),
-					file: posixified
-				});
-			}
+			remotes.push({
+				hash: hash(posixified),
+				file: posixified
+			});
 		}
 	}