+++ packages/kit/src/exports/vite/index.js
@@ -15,9 +15,8 @@ import { build_server_nodes } from './build/build_server.js';
 import { build_service_worker } from './build/build_service_worker.js';
 import { assets_base, find_deps, resolve_symlinks } from './build/utils.js';
 import { dev } from './dev/index.js';
-import { is_illegal, module_guard } from './graph_analysis/index.js';
 import { preview } from './preview/index.js';
-import { get_config_aliases, get_env, normalize_id, strip_virtual_prefix } from './utils.js';
+import { get_config_aliases, get_env, normalize_id, stackless } from './utils.js';
 import { write_client_manifest } from '../../core/sync/write_client_manifest.js';
 import prerender from '../../core/postbuild/prerender.js';
 import analyse from '../../core/postbuild/analyse.js';
@@ -37,6 +36,7 @@ import {
 import { import_peer } from '../../utils/import.js';
 import { compact, conjoin } from '../../utils/array.js';
 import { build_remotes, treeshake_prerendered_remotes } from './build/build_remote.js';
+import { should_ignore } from './static_analysis/utils.js';
 
 const cwd = process.cwd();
 
@@ -90,7 +90,7 @@ const warning_preprocessor = {
 		const basename = path.basename(filename);
 		if (basename.startsWith('+page.') || basename.startsWith('+layout.')) {
 			const match = content.match(options_regex);
-			if (match) {
+			if (match && match.index !== undefined && !should_ignore(content, match.index)) {
 				const fixed = basename.replace('.svelte', '(.server).js/ts');
 
 				const message =
@@ -219,6 +219,7 @@ async function kit({ svelte_config }) {
 
 	const normalized_cwd = vite.normalizePath(cwd);
 	const normalized_lib = vite.normalizePath(kit.files.lib);
+	const normalized_node_modules = vite.normalizePath(path.resolve('node_modules'));
 
 	/**
 	 * A map showing which features (such as `$app/server:read`) are defined
@@ -374,9 +375,6 @@ async function kit({ svelte_config }) {
 		}
 	};
 
-	/** @type {Map<string, string>} */
-	const import_map = new Map();
-
 	/** @type {import('vite').Plugin} */
 	const plugin_virtual_modules = {
 		name: 'vite-plugin-sveltekit-virtual-modules',
@@ -389,6 +387,7 @@ async function kit({ svelte_config }) {
 			// If importing from a service-worker, only allow $service-worker & $env/static/public, but none of the other virtual modules.
 			// This check won't catch transitive imports, but it will warn when the import comes from a service-worker directly.
 			// Transitive imports will be caught during the build.
+			// TODO move this logic to plugin_guard
 			if (importer) {
 				const parsed_importer = path.parse(importer);
 
@@ -405,8 +404,6 @@ async function kit({ svelte_config }) {
 						)} into service-worker code. Only the modules $service-worker and $env/static/public are available in service workers.`
 					);
 				}
-
-				import_map.set(id, importer);
 			}
 
 			// treat $env/static/[public|private] as virtual
@@ -414,9 +411,11 @@ async function kit({ svelte_config }) {
 				// ids with :$ don't work with reverse proxies like nginx
 				return `\0virtual:${id.substring(1)}`;
 			}
+
 			if (id === '__sveltekit/remote') {
 				return `${runtime_directory}/client/remote-functions/index.js`;
 			}
+
 			if (id.startsWith('__sveltekit/')) {
 				return `\0virtual:${id}`;
 			}
@@ -428,37 +427,6 @@ async function kit({ svelte_config }) {
 			const global = is_build
 				? `globalThis.__sveltekit_${version_hash}`
 				: 'globalThis.__sveltekit_dev';
-
-			if (options?.ssr === false && process.env.TEST !== 'true') {
-				if (
-					is_illegal(id, {
-						cwd: normalized_cwd,
-						node_modules: vite.normalizePath(path.resolve('node_modules')),
-						server: vite.normalizePath(path.join(normalized_lib, 'server'))
-					})
-				) {
-					const relative = normalize_id(id, normalized_lib, normalized_cwd);
-
-					const illegal_module = strip_virtual_prefix(relative);
-
-					const error_prefix = `Cannot import ${illegal_module} into client-side code. This could leak sensitive information.`;
-					const error_suffix = `
-Tips:
- - To resolve this error, ensure that no exports from ${illegal_module} are used, even transitively, in client-side code.
- - If you're only using the import as a type, change it to \`import type\`.
- - If you're not sure which module is causing this, try building your app -- it will create a more helpful error.`;
-
-					if (import_map.has(illegal_module)) {
-						const importer = path.relative(
-							cwd,
-							/** @type {string} */ (import_map.get(illegal_module))
-						);
-						throw new Error(`${error_prefix}\nImported by: ${importer}.${error_suffix}`);
-					}
-
-					throw new Error(`${error_prefix}${error_suffix}`);
-				}
-			}
 
 			switch (id) {
 				case env_static_private:
@@ -566,6 +534,10 @@ Tips:
 		}
 	};
 
+	/** @type {Map<string, Set<string>>} */
+	const import_map = new Map();
+	const server_only_pattern = /.*\.server\..+/;
+
 	/**
 	 * Ensures that client-side code can't accidentally import server-side code,
 	 * whether in `*.server.js` files, `$app/server`, `$lib/server`, or `$env/[static|dynamic]/private`
@@ -574,23 +546,89 @@ Tips:
 	const plugin_guard = {
 		name: 'vite-plugin-sveltekit-guard',
 
-		writeBundle: {
-			sequential: true,
-			handler(_options) {
-				if (vite_config.build.ssr) return;
+		// Run this plugin before built-in resolution, so that relative imports
+		// are added to the module graph
+		enforce: 'pre',
 
-				const guard = module_guard(this, {
-					cwd: vite.normalizePath(process.cwd()),
-					lib: vite.normalizePath(kit.files.lib)
-				});
+		async resolveId(id, importer) {
+			if (importer && !importer.endsWith('index.html')) {
+				const resolved = await this.resolve(id, importer, { skipSelf: true });
 
-				manifest_data.nodes.forEach((_node, i) => {
-					const id = vite.normalizePath(
-						path.resolve(kit.outDir, `generated/client-optimized/nodes/${i}.js`)
-					);
+				if (resolved) {
+					const normalized = normalize_id(resolved.id, normalized_lib, normalized_cwd);
 
-					guard.check(id);
-				});
+					let importers = import_map.get(normalized);
+
+					if (!importers) {
+						importers = new Set();
+						import_map.set(normalized, importers);
+					}
+
+					importers.add(normalize_id(importer, normalized_lib, normalized_cwd));
+				}
+			}
+		},
+
+		load(id, options) {
+			if (options?.ssr === true || process.env.TEST === 'true') {
+				return;
+			}
+
+			// skip .server.js files outside the cwd or in node_modules, as the filename might not mean 'server-only module' in this context
+			const is_internal = id.startsWith(normalized_cwd) && !id.startsWith(normalized_node_modules);
+
+			const normalized = normalize_id(id, normalized_lib, normalized_cwd);
+
+			const is_server_only =
+				normalized === '$env/static/private' ||
+				normalized === '$env/dynamic/private' ||
+				normalized === '$app/server' ||
+				normalized.startsWith('$lib/server/') ||
+				(is_internal && server_only_pattern.test(path.basename(id)));
+
+			if (is_server_only) {
+				// in dev, this doesn't exist, so we need to create it
+				manifest_data ??= sync.all(svelte_config, vite_config_env.mode).manifest_data;
+
+				/** @type {Set<string>} */
+				const entrypoints = new Set();
+				for (const node of manifest_data.nodes) {
+					if (node.component) entrypoints.add(node.component);
+					if (node.universal) entrypoints.add(node.universal);
+				}
+
+				const normalized = normalize_id(id, normalized_lib, normalized_cwd);
+				const chain = [normalized];
+
+				let current = normalized;
+
+				while (true) {
+					const importers = import_map.get(current);
+					if (!importers) break;
+
+					const candidates = Array.from(importers).filter((importer) => !chain.includes(importer));
+					if (candidates.length === 0) break;
+
+					chain.push((current = candidates[0]));
+
+					if (entrypoints.has(current)) {
+						let message = `Cannot import ${normalized} into code that runs in the browser, as this could leak sensitive information.`;
+
+						const pyramid = chain
+							.reverse()
+							.map((id, i) => {
+								return `${' '.repeat(i + 1)}${id}`;
+							})
+							.join(' imports\n');
+
+						message += `\n\n${pyramid}`;
+						message += `\n\nIf you're only using the import as a type, change it to \`import type\`.`;
+
+						throw stackless(message);
+					}
+				}
+
+				throw new Error('An impossible situation occurred');
 			}
 		}
 	};
@@ -974,23 +1012,36 @@ Tips:
 
 				secondary_build_started = true;
 
-				const { output: client_chunks } = /** @type {import('vite').Rollup.RollupOutput} */ (
-					await vite.build({
-						configFile: vite_config.configFile,
-						// CLI args
-						mode: vite_config_env.mode,
-						logLevel: vite_config.logLevel,
-						clearScreen: vite_config.clearScreen,
-						build: {
-							minify: initial_config.build?.minify,
-							assetsInlineLimit: vite_config.build.assetsInlineLimit,
-							sourcemap: vite_config.build.sourcemap
-						},
-						optimizeDeps: {
-							force: vite_config.optimizeDeps.force
-						}
-					})
-				);
+				let client_chunks;
+
+				try {
+					const bundle = /** @type {import('vite').Rollup.RollupOutput} */ (
+						await vite.build({
+							configFile: vite_config.configFile,
+							// CLI args
+							mode: vite_config_env.mode,
+							logLevel: vite_config.logLevel,
+							clearScreen: vite_config.clearScreen,
+							build: {
+								minify: initial_config.build?.minify,
+								assetsInlineLimit: vite_config.build.assetsInlineLimit,
+								sourcemap: vite_config.build.sourcemap
+							},
+							optimizeDeps: {
+								force: vite_config.optimizeDeps.force
+							}
+						})
+					);
+
+					client_chunks = bundle.output;
+				} catch (e) {
+					const error =
+						e instanceof Error ? e : new Error(/** @type {any} */ (e).message ?? e ?? '<unknown>');
+
+					// without this, errors that occur during the secondary build
+					// will be logged twice
+					throw stackless(error.stack ?? error.message);
+				}
 
 				copy(
 					`${out}/server/${kit.appDir}/immutable/assets`,