+++ packages/kit/test/prerendering/paths-base/src/routes/assets/emitted/+page.server.js
@@ -1,0 +1,8 @@
+// @ts-ignore
+import url from './message.csv?url';
+
+export async function load({ fetch }) {
+	const response = await fetch(url);
+	const message = await response.text();
+	return { message };
+}